+++ source_files/edge/rad_trig.h
@@ -49,7 +49,7 @@ bool              ScriptRadiusCheck(MapObject *mo, RADScript *r);
 RADScript        *FindScriptByName(const char *map_name, const char *name);
 RADScriptTrigger *FindScriptTriggerByName(const char *name);
 RADScriptState   *FindScriptStateByLabel(RADScript *scr, char *label);
-void              ScriptEnableByTag(MapObject *actor, uint32_t tag, bool disable, RADScriptTag tagtype);
+void              ScriptEnableByTag(MapObject *actor, uint64_t tag, bool disable, RADScriptTag tagtype);
 void              ScriptEnableByTag(MapObject *actor, const char *name, bool disable);
 bool              CheckActiveScriptByTag(MapObject *actor, const char *name);
 void              ScriptUpdateMonsterDeaths(MapObject *mo);
